Logout and restricted pages

+++ base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
@@ -15,6 +17,10 @@ from .forms import RoomForm
 ]
 '''
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST': 
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -29,20 +35,15 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, 'Username or password does not exist ')
-            
-
-    
-
-
-    
-
-
-
-
-
 
     context = {}
     return render(request, 'base/login_register.html', context)
+
+
+
+def logoutUser(request):
+    logout(request)
+    return(redirect('home'))
 
 
 
@@ -70,7 +71,7 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -82,10 +83,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not available')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -96,9 +101,11 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context )
 
-
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Not available')
     if request.method == 'POST':
         room.delete()
         return redirect ('home')
